restructured files and connected servers
