+++ controllers/users.js
@@ -1,0 +1,12 @@
+// Link to home routes
+
+import usersModel from "../models/users";
+
+const userController = {
+  getUsers: (req, res, next) => {
+    const users = usersModel.getUsers()
+    res.send(users);
+  }
+}
+
+export default userController
