--- expense-tracker/.env.development
@@ -1,1 +1,0 @@
-PORT=8081
