--- expense-tracker/.express/init/dotenv.js
@@ -1,6 +1,0 @@
-import path from "path"
-import dotenv from "dotenv"
-import { dirname } from "dirname-filename-esm"
-import expressConfig from "../../express.config"
-
-dotenv.config({ path: path.join(dirname(import.meta), "../../", `${expressConfig.env[process.env.NODE_ENV]}`) })
