--- expense-tracker/.express/init/index.js
@@ -1,3 +1,0 @@
-import "./dotenv"
-import "./server"
-import "./welcome"
