--- expense-tracker/.express/init/server.js
@@ -1,22 +1,0 @@
-
-export const PORT = normalizePort(process.env.PORT || '3000')
-
-/**
- * Normalize a port into a number, string, or false.
- */
-
-function normalizePort(val) {
-  var port = parseInt(val, 10);
-
-  if (isNaN(port)) {
-    // named pipe
-    return val;
-  }
-
-  if (port >= 0) {
-    // port number
-    return port;
-  }
-
-  return false;
-}
