--- expense-tracker/.express/init/welcome.js
@@ -1,8 +1,0 @@
-import { PORT } from "./server"
-
-console.log("")
-// log node-env
-console.log("- NODE_ENV=" + process.env.NODE_ENV)
-// log port
-console.log("- PORT=" + PORT)
-console.log("")
