--- expense-tracker/.express/www.js
@@ -1,73 +1,0 @@
-#!/usr/bin/env node
-
-import "./init/index"
-
-/**
- * Module dependencies.
- */
-
-import app from '../src/app';
-import debug from 'debug';
-import http from 'http';
-import { PORT } from "./init/server";
-const serverDebug = debug('server:server')
-
-/**
- * Get port from environment and store in Express.
- */
-
-app.set('port', PORT);
-
-/**
- * Create HTTP server.
- */
-
-var server = http.createServer(app);
-
-/**
- * Listen on provided port, on all network interfaces.
- */
-
-server.listen(PORT);
-server.on('error', onError);
-server.on('listening', onListening);
-
-/**
- * Event listener for HTTP server "error" event.
- */
-
-function onError(error) {
-  if (error.syscall !== 'listen') {
-    throw error;
-  }
-
-  var bind = typeof PORT === 'string'
-    ? 'Pipe ' + PORT
-    : 'Port ' + PORT;
-
-  // handle specific listen errors with friendly messages
-  switch (error.code) {
-    case 'EACCES':
-      console.error(bind + ' requires elevated privileges');
-      process.exit(1);
-      break;
-    case 'EADDRINUSE':
-      console.error(bind + ' is already in use');
-      process.exit(1);
-      break;
-    default:
-      throw error;
-  }
-}
-
-/**
- * Event listener for HTTP server "listening" event.
- */
-
-function onListening() {
-  var addr = server.address();
-  var bind = typeof addr === 'string'
-    ? 'pipe ' + addr
-    : 'port ' + addr.port;
-  serverDebug('Listening on ' + bind);
-}
