--- expense-tracker/README.md
@@ -1,27 +1,0 @@
-# Getting Started with Gen Express App
-
-This project was bootstrapped with [Gen Express App](https://github.com/Dalufishe/gen-express-app).
-
-## Available Scripts
-
-In the project directory, you can run:
-
-### `npm run dev`
-
-Runs the app in the development mode.\
-Open [http://localhost:8081](http://localhost:8081) to view it in your browser.
-
-The port can be modified in the .env.development file.
-This mode includes a hot reloader.
-
-### `npm run start`
-
-Runs the app in the development mode.\
-Open [http://localhost:3000](http://localhost:3000) to view it in your browser.
-
-The port can be modified in the .env file.
-This mode utilizes an optimized build.
-
-## Learn More
-
-You can learn more in the [Gen Express App Github README](https://github.com/Dalufishe/gen-express-app).
