--- expense-tracker/express.config.js
@@ -1,6 +1,0 @@
-export default {
-  env: {
-    production: ".env",
-    development: ".env.development",
-  },
-};
