--- expense-tracker/public/stylesheets/style.css
@@ -1,9 +1,0 @@
-
-body {
-  padding: 50px;
-  font: 14px "Lucida Grande", Helvetica, Arial, sans-serif;
-}
-
-a {
-  color: #00b7ff;
-}
