--- expense-tracker/src/controllers/users.js
@@ -1,12 +1,0 @@
-// Link to home routes
-
-import usersModel from "../models/users";
-
-const userController = {
-  getUsers: (req, res, next) => {
-    const users = usersModel.getUsers()
-    res.send(users);
-  }
-}
-
-export default userController
