--- expense-tracker/src/routers/users.js
@@ -1,8 +1,0 @@
-import express from 'express';
-const router = express.Router();
-
-import userController from "../controllers/users";
-
-router.get('/', userController.getUsers);
-
-export default router;
